+++ pycm/pycm_param.py
@@ -431,7 +431,8 @@ PARAMS_LINK = {
     "AUPR": "AUPR-(Area-under-the-PR-curve)",
     "ICSI": "ICSI-(Individual-classification-success-index)",
     "CSI": "CSI-(Classification-success-index)",
-    "QI": "QI-(Yule's-Q-interpretation)"}
+    "QI": "QI-(Yule's-Q-interpretation)",
+    "ARI": "ARI-(Adjusted-Rand-index)"}
 
 CAPITALIZE_FILTER = ["BCD", "AUCI", "Q", "AGF", "OOC", "AUPR", "AUC", "QI"]